phew finally running tests
