+++ autofit/graphical/__init__.py
@@ -1,24 +1,19 @@
-from .declarative.factor.analysis import AnalysisFactor
+from . import optimise as optimize
+from . import utils
 from .declarative.collection import FactorGraphModel
+from .declarative.factor.analysis import AnalysisFactor
 from .declarative.factor.hierarchical import HierarchicalFactor
-
+from .expectation_propagation import EPMeanField, EPOptimiser
 from .factor_graphs import (
-    Factor, 
-    FactorJacobian, 
-    FactorGraph, 
-    AbstractFactor, 
+    Factor,
+    FactorJacobian,
+    FactorGraph,
+    AbstractFactor,
     FactorValue,
     VariableTransform,
     FullCholeskyTransform,
     identity_transform
 )
 from .mean_field import FactorApproximation, MeanField
-from .expectation_propagation import EPMeanField, EPOptimiser
-from .messages import FixedMessage, NormalMessage, UniformNormalMessage, GammaMessage, AbstractMessage, BetaMessage, UniformNormalMessage, LogNormalMessage, MultiLogitNormalMessage
-
 from .optimise import OptFactor, LaplaceFactorOptimiser, lstsq_laplace_factor_approx
 from .sampling import ImportanceSampler, project_factor_approx_sample
-from ..mapper.variable import Variable, Plate
-
-from . import optimise as optimize
-from . import utils
