+++ test_autofit/graphical/functionality/test_projection.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 from scipy import integrate
 
-import autofit.graphical.messages.normal
+import autofit.messages.normal
 from autofit import graphical as mp
 
 
@@ -10,7 +10,7 @@ from autofit import graphical as mp
     name="q_cavity"
 )
 def make_q_cavity():
-    return autofit.graphical.messages.normal.NormalMessage(-0.5, 0.5)
+    return autofit.messages.normal.NormalMessage(-0.5, 0.5)
 
 
 def test_integration(
@@ -32,7 +32,7 @@ def test_integration(
     ni_0, ni_1, ni_2 = (
         integrate.trapz(x ** i * tilted_distribution, x) for i in range(3))
 
-    q_numerical = autofit.graphical.messages.normal.NormalMessage.from_sufficient_statistics(
+    q_numerical = autofit.messages.normal.NormalMessage.from_sufficient_statistics(
         [ni_1 / ni_0, ni_2 / ni_0]
     )
 
@@ -68,7 +68,7 @@ def test_laplace_method(probit_factor, q_cavity, x):
     opt_probit = mp.OptFactor.from_approx(probit_approx)
     result = opt_probit.maximise({x: 0.})
 
-    q_probit_laplace = autofit.graphical.messages.normal.NormalMessage.from_mode(
+    q_probit_laplace = autofit.messages.normal.NormalMessage.from_mode(
         result.mode[x],
         covariance=result.hess_inv[x]
     )
